requirements.txt
pytest
selenium
webdriver-manager
pytest-pyreport
allure-pytest
